Use db.item as fallback
Adds a fallback to handle if the item isn't present in db.templates for some reason (newly added to game or added by mod). If such an item is checked for insurance (or insured) it will use the items database price instead of it's template price. This closes #415 but the relevant function getItemPrice() should be moved to a more appropriate location later in development.