Fixing Insurance Exceptions
insuredItem is only an itemId. As such the check for slotId was failing to catch the relevant items. Then the parenting code (now at Ln59) would change their slotId which removes the protections that those items have from being removed in offraid.

The loop is a bodge to protect the items within the secure container from having their insurance applied/removed. Will relocate this loop to helpers eventually, but this resolves the issues for now.